// daily-app-frontend/Daily/src/app/validation/models/aggregation-data.model.ts
export interface AggregationData {
  dateKey?: any;
  fullDate?: string;
  inSubs?: number;
  outSubs?: number;
  ratePlan?: string;
  ratePlanKey?: number;
  opening?: number;
  closing?: number;
  dwhStatus?: string;
  dwhStatusKey?: number;
  transactionTypeKey?: any;
  inAdjustFlag?: number;
  outAdjustFlag?: number;
  date?: string;
  dayDateKey?: any;
  inSubTransactionTypeKey?: any;
  outSubTransactionTypeKey?: any;
}