// daily-app-frontend/Daily/src/app/manual-adjustment/models/manual-adjustment.model.ts
export interface ManualAdjustment {
  closing?: number;
  dateKey?: number;
  opening?: number;
  ratePlan?: string;
  ratePlanGroupKey?: number;
  ratePlanKey?: number;
  ratePlanType?: number;
  variance?: number;
  numberOfSubs?: number;
  dwhStatusKey?: number;
  trxTypeKey?: number;
  dayDateKey?: number;
}